worker/requirements.txt
pika==1.3.2
minio==7.2.9
psycopg2-binary==2.9.9
httpx==0.27.0
